Extract `FixtureSupport` trait from `MyBenchmark`.
This makes it easier to reuse it from JMH-based benchmarks.
