+++ benchmark/src/main/scala/spire/benchmark/Benchmark.scala
@@ -22,34 +22,7 @@ import java.math.BigInteger
 /**
  * Extend this to create an actual benchmarking class.
  */
-trait MyBenchmark extends SimpleBenchmark {
-
-  /**
-   * Sugar for building arrays using a per-cell init function.
-   */
-  def init[A:ClassTag](size:Int)(f: => A) = {
-    val data = Array.ofDim[A](size)
-    for (i <- 0 until size) data(i) = f
-    data
-  }
-
-  /**
-   * Sugar for building arrays using a per-cell init function.
-   */
-  def mkarray[A:ClassTag:Order](size:Int, layout:String)(f: => A): Array[A] = {
-    val data = init(size)(f)
-    val ct = implicitly[ClassTag[A]]
-    val order = Order[A]
-    layout match {
-      case "random" =>
-      case "sorted" => spire.math.Sorting.sort(data)(order, ct)
-      case "reversed" => spire.math.Sorting.sort(data)(order.reverse, ct)
-      case _ => sys.error(s"unknown layout: $layout")
-    }
-    data
-  }
-
-  def nextComplex = Complex(nextDouble, nextDouble)
+trait MyBenchmark extends SimpleBenchmark with FixtureSupport {
 
   /**
    * Sugar to run 'f' for 'reps' number of times.
